+++ packages/veritone-widgets/CHANGELOG.md
@@ -204,3 +204,7 @@ const MyPage = () => (
 
 ## 7.1.3
 * Namespace global styles used by media player widget
+
+## 7.2.0
+* New multiple engine picker and selected engines info panel
+* Beta: InfinitePicker exported as part of engine picker.